@@ -2,6 +2,7 @@
   "name": "too",
   "version": "2.4.0",
   "description": "Combine multiple processes' Stdout/Stderr/SIGINT to keep them all foreground",
+  "type": "module",
   "scripts": {
     "build": "tsc",
     "validate": "npm run lint && npm run build && npm link && too -c 'echo foo' -c 'sleep 5' -c 'gcloud beta emulators datastore start --project pricealert-jp' -c 'echo bar'",
@@ -27,14 +28,17 @@
     "lookpath": "^1.2.2"
   },
   "devDependencies": {
+    "@eslint/js": "^9.12.0",
     "@types/jest": "^29.5.12",
     "@types/node": "^22.0.0",
-    "@typescript-eslint/eslint-plugin": "^4.15.2",
-    "@typescript-eslint/parser": "^4.15.2",
-    "eslint": "^7.20.0",
+    "@typescript-eslint/eslint-plugin": "^8.8.1",
+    "@typescript-eslint/parser": "^8.8.1",
+    "eslint": "^9.12.0",
+    "globals": "^15.11.0",
     "jest": "^29.7.0",
     "ts-jest": "^29.2.3",
-    "typescript": "^5.6.2"
+    "typescript": "^5.6.2",
+    "typescript-eslint": "^8.8.1"
   },
   "jest": {
     "collectCoverage": true,
